refactor: combine drawBorder() with clearScreen()
Because every call drawBorder() does a clearScreen() beforehand,
we can combine both functions.

With this knowledge we can also restrict the Fill() call to the space
inside the border - no need to write the border map tiles twice.

This gives us a for a handy reusable drawBox() method that I want to
reuse for a better game over screen later.