add left and right borders to game area
At least with my Uzebox and my Sony flatscreen some of the screen is
cut off on the left and right.  Add some borders to make the whole
level visible.

With the recently introduces borders, the level got 1 row and 6
columns smaller.  I hope this does not affect the gameplay too much.