refactor: remove tmp_level optimization
`tmp_level` was a manual optimization trying to reduce repeated
accesses to the same memory location when testing tiles.

But: don't optimize without measuring!

The overhead of initializing and updating `tmp_level` was not taken
into account.  By removing the variable, GCC was able to save 60 bytes
of code.  It's clearly better at optimizing than me.