possible fix for worm erratics and lockups
Sometimes worms that run off the bottom spawn very strange at the top,
the game might lock up shortly after.  Holes in the level borders to
the sides have been observed.

I suspect a possible problem with a worm spawning right next to the
border combined with mushrooms in the immediate vincinity and the
following direction changes.

Imaginary solution: don't respawn next to the borders.