fix: reset worm body positions after headshot
`shootWormBody()` does a reverse lookup via `wormx[]` and `wormy[]` to
find the index that got shot.  Next this index is mapped to a current
worm.

When `wormx[]`/`wormy[]` were not cleaned up by `shootWormHead()`
after a whole worm was shot AND later `shootWormBody()` checked
another worm shot AT THE SAME LOCATION, the reverse lookup would
sometimes find a matching leftover `wormx`/`wormy` from the already
killed older worm.  This left us with an index that was not assigned
to any current worm, which made the next lookup (index -> worm) fail
in weird ways.

For the same reason we now always reset all `wormx[]` and `wormy[]`
when initializing worms instead of only the parts that the new worms
use.