fix possible bug when a worm can't be split
When the game configuration is changed so that MAXWORMCOUNT <
MAXWORMLEN - 1, a problem can occur:

When a worm is split, there might be no free worm index available to
hold the new worm that spawns from the tail of the old worm.  Catch
this case and instead of spawning a new worm just convert the tail to
mushrooms.

In normal game configuration where MAXWORMCOUNT == MAXWORMLEN,
this can never happen.

Because the body part shot is converted to a mushroom, it can't even
happen at MAXWORMCOUNT == MAXWORMLEN - 1.