fix shootWormBody(): split idx must also be set to OFFSCREEN
We have already fixed setting wormx[]/wormy[] to OFFSCREEN for
whole-worm kills and the special case "kill whole worm after the split
point because we can't create a new worm".

But until now we have missed to also set the split index (where the
worm body was shot) to OFFSCREEN.

All hail the new debug output!  It immediately showed the problem.
I was very lucky to trigger this corner case:
- I shot a worm at exactly the same position where I have split
  a worm before.
- The old split was still in memory (no "all worms killed" inbetween).
- The old split had a lower wormx[]/wormy[] index than the currently
  shot worm.