fix bug: second part of split worm could be spawned offscreen
If a worm body gets a shot exactly atthe worm's spawn point,
the part after the split will still be offscreen.

If we make a new worm out of the part still offscreen,
the new worm will also be completely offscreen and
the next movement of that worm's head will crash.

To fix this, the second part of the worm will not be turned into a new
worm if its first body element is offscreen.  The split point will
turn in to a mushroom as usual, but the worm will not be split.  The
offscreen part will be marked free and can be reused later.