split offscreen constant
To hunt down a bug, the offscreen constant OFFSCREEN that was used
both as an X and Y coordinate has now been split into

* OFFSCREEN_X, a generic "this thing is currently offscreen"

* OFFSCREEN_Y_*, showing what exactly is offscreen

I mainly want to split "worm offscreen" into "worm part is currently
offscreen and belongs to a worm" and "worm part is currently offscren
and does not belong to a worm".

Only completely free parts can be used when expanding a worm that runs
off the bottom of the screen.  I suspect a bug somewhere in the old
implementation and want to make it visible.