extract duplicate code for offscreen setting
Also include "shooting = false" in setShotOffscreen()
and always call setShotOffscreen() where shooting was
set to false before.

There seem to have been instances where the shot was set to false, but
the shot was not set offscreen and the monsters could run into an
invisible already removed shot and die.
This error should be fixed now.